all work, fix bugs and working view

+++ sevices/films.js
@@ -23,13 +23,18 @@ class FilmsService extends CrudService {
     }
     async delete(id) {
         const film = await super.read(id);
-        fs.unlink(pathFolder + film.preview_path, (err) => {});
+
+        if(isEmpty(film.preview_path))
+            fs.unlink(pathFolder + film.preview_path, (err) => {});
 
         return await super.repository.destroy({ where: { id } });
     }
     async update(id, data) {
         const film = await super.read(id);
-        fs.unlink(pathFolder + film.preview_path, (err) => {});
+
+        if(isEmpty(film.preview_path))
+            fs.unlink(pathFolder + film.preview_path, (err) => {});
+
         await super.repository.update(data, { where: { id }, limit: 1 });
         
 		return await this.read(id);
