+++ sevices/videos.js
@@ -34,13 +34,17 @@ class VideosService extends CrudService {
     }
     async delete(id) {
         const video = await super.read(id);
-        fs.unlink(pathFolder + video.preview_path, (err) => {});
+        if(isEmpty(video.preview_path))
+            fs.unlink(pathFolder + video.preview_path, (err) => {});
 
         return await super.repository.destroy({ where: { id } });
     }
     async update(id, data) {
         const video = await super.read(id);
-        fs.unlink(pathFolder + video.preview_path, (err) => {});
+
+        if(isEmpty(video.preview_path))
+            fs.unlink(pathFolder + video.preview_path, (err) => {});
+            
         await super.repository.update(data, { where: { id }, limit: 1 });
 
 		return await this.read(id);
